Refactor code before PR
