--- connectors/uplynk/src/main/java/com/theoplayer/android/connector/uplynk/api/PreplayResponse.kt
@@ -1,8 +1,0 @@
-package com.theoplayer.android.connector.uplynk.api
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-class PreplayResponse(val playURL: String) {
-
-}
