+++ connectors/uplynk/src/main/java/com/theoplayer/android/connector/uplynk/network/PreplayResponse.kt
@@ -1,0 +1,6 @@
+package com.theoplayer.android.connector.uplynk.network
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+internal class PreplayResponse(val playURL: String)
