+++ connectors/uplynk/src/main/java/com/theoplayer/android/connector/uplynk/network/UplynkApi.kt
@@ -1,0 +1,14 @@
+package com.theoplayer.android.connector.uplynk.network
+
+import android.util.Log
+import kotlinx.serialization.json.Json
+
+internal class UplynkApi {
+    private val json = Json{ ignoreUnknownKeys = true }
+    private val network = HttpsConnection()
+
+    suspend fun preplay(srcURL: String): PreplayResponse {
+        val body = network.get(srcURL)
+        return json.decodeFromString(body)
+    }
+}
